+++ provider/src/main/java/com/raygun/raygun4android/RUM.java
@@ -19,6 +19,7 @@ import java.util.UUID;
 import java.util.concurrent.TimeUnit;
 
 import com.google.gson.Gson;
+import com.raygun.raygun4android.logging.RaygunLogger;
 import com.raygun.raygun4android.messages.rum.RaygunRUMData;
 import com.raygun.raygun4android.messages.rum.RaygunRUMDataMessage;
 import com.raygun.raygun4android.messages.rum.RaygunRUMMessage;
@@ -39,7 +40,7 @@ public class RUM implements ActivityLifecycleCallbacks {
 
     private static void attach(Activity mainActivity) {
 
-        RaygunLogger.v("attach");
+        RaygunLogger.v("attached RUM");
         if (RUM.rum == null && mainActivity != null) {
             Application application = mainActivity.getApplication();
 
@@ -105,7 +106,7 @@ public class RUM implements ActivityLifecycleCallbacks {
 
     @Override
     public void onActivityCreated(Activity activity, Bundle bundle) {
-        RaygunLogger.v("onActivityCreated");
+        RaygunLogger.v("RUM - onActivityCreated");
         if (RUM.currentActivity == null) {
             if (doesNeedSessionRotation()) {
                 rotateSession(currentSessionUser,currentSessionUser);
@@ -123,7 +124,7 @@ public class RUM implements ActivityLifecycleCallbacks {
 
     @Override
     public void onActivityStarted(Activity activity) {
-        RaygunLogger.v("onActivityStarted");
+        RaygunLogger.v("RUM - onActivityStarted");
         if (RUM.currentActivity == null) {
             if (doesNeedSessionRotation()) {
                 rotateSession(currentSessionUser,currentSessionUser);
@@ -141,7 +142,7 @@ public class RUM implements ActivityLifecycleCallbacks {
 
     @Override
     public void onActivityResumed(Activity activity) {
-        RaygunLogger.v("onActivityResumed");
+        RaygunLogger.v("RUM - onActivityResumed");
         if (RUM.currentActivity == null) {
             if (doesNeedSessionRotation()) {
                 rotateSession(currentSessionUser,currentSessionUser);
@@ -166,13 +167,13 @@ public class RUM implements ActivityLifecycleCallbacks {
 
     @Override
     public void onActivityPaused(Activity activity) {
-        RaygunLogger.v("onActivityPaused");
+        RaygunLogger.v("RUM - onActivityPaused");
         RUM.lastSeenTime = System.currentTimeMillis();
     }
 
     @Override
     public void onActivityStopped(Activity activity) {
-        RaygunLogger.v("onActivityStopped");
+        RaygunLogger.v("RUM - onActivityStopped");
         if (activity == RUM.currentActivity) {
             RUM.currentActivity = null;
             RUM.loadingActivity = null;
@@ -182,7 +183,7 @@ public class RUM implements ActivityLifecycleCallbacks {
 
     @Override
     public void onActivitySaveInstanceState(Activity activity, Bundle bundle) {
-        RaygunLogger.v("onActivitySIS");
+        RaygunLogger.v("RUM - onActivitySIS");
         RUM.lastSeenTime = System.currentTimeMillis();
     }