+++ provider/src/main/java/com/raygun/raygun4android/RaygunClient.java
@@ -7,6 +7,8 @@ import android.content.pm.ApplicationInfo;
 import android.content.pm.PackageManager;
 import android.os.Bundle;
 
+import com.raygun.raygun4android.logging.RaygunLogger;
+import com.raygun.raygun4android.logging.TimberRaygunLoggerImplementation;
 import com.raygun.raygun4android.messages.crashreporting.RaygunBreadcrumbMessage;
 import com.raygun.raygun4android.messages.shared.RaygunUserInfo;
 import com.raygun.raygun4android.utils.RaygunFileUtils;
@@ -62,6 +64,8 @@ public class RaygunClient {
      * @param version The version of your application, format x.x.x.x, where x is a positive integer.
      */
     public static void init(Application application, String apiKey, String version) {
+
+        TimberRaygunLoggerImplementation.init();
 
         RaygunLogger.d("Configuring Raygun4Android (v" + RaygunSettings.RAYGUN_CLIENT_VERSION + ")");